requirements.txt
attrs==25.4.0
beautifulsoup4==4.14.3
certifi==2025.11.12
chardet==5.2.0
charset-normalizer==3.4.4
h11==0.16.0
idna==3.11
lxml==6.0.2
numpy==2.4.1
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
selenium==4.39.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.1
trio==0.32.0
trio-websocket==0.12.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
webdriver-manager==4.0.2
websocket-client==1.9.0
wsproto==1.3.2
